Moved field_class div within if statement (#35)
Bootstrap5 has limited use for field_class as the labels and divs are now siblings for each other. wrapping_class is on the div wrapping both the label and input is more likely to be used. This change is needed as the new bs5 floating labels need to be siblings.