Move radio checkbox select help text out of div (#123)
Currently, the help text is included in the div, if a scroll class is added to `field_class`,
it will result in the errors and help text being shown as part of the scroll and have to
scroll to the bottom of the inner container to see it. This likely won't be surprising as
the other fields, the help text are beside the input and not within the input container.